Move contact-related parsers into package contact
