+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -32,6 +32,7 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.contact.AddCommand;
+import seedu.address.logic.parser.contact.AddCommandParser;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
