+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -9,6 +9,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.contact.FindCommand;
+import seedu.address.logic.parser.contact.FindCommandParser;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
